% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_extract.r
\name{extract_formulas}
\alias{extract_formulas}
\title{Extract formulas from an Excel workbook sheet}
\usage{
extract_formulas(xlsx, n)
}
\arguments{
\item{xlsx}{an \code{xlsx} object readin with \code{read_xlsx()}}

\item{n}{sheet number}
}
\value{
a \code{tibble} with three columns (sheet, cell & formuma text). The
    data frame may be empty if no formulas were found. The return will be \code{NULL}
        if the sheet wasn't found and a warning to said effect will be issued.
}
\description{
Extract formulas from an Excel workbook sheet
}
\examples{
doc <- read_xlsx(system.file("extdata/wb.xlsx", package="xlsxtractr"))

extract_formulas(doc, 1)
extract_formulas(doc, 2)
extract_formulas(doc, 3) # no formula

# gotta extract'em all!
map_df(seq_along(doc), ~extract_formulas(doc, .))
}

