% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_extract.r
\name{read_xlsx}
\alias{read_xlsx}
\title{Read in an Excel document for various extractions}
\usage{
read_xlsx(path)
}
\arguments{
\item{path}{path to the Excel (xlsx) document}
}
\description{
Local file path or URL pointing to a \code{.xlsx} file.
}
\examples{
doc <- read_xlsx(system.file("extdata/wb.xlsx", package="xlsxtractr"))
class(doc)
}

