% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxtractr-package.R
\docType{package}
\name{xlsxtractr}
\alias{xlsxtractr}
\alias{xlsxtractr-package}
\title{Extract Things From Excel (xlsx) Files}
\description{
Extract Things From Excel (xlsx) Files
}
\author{
Bob Rudis (@hrbrmstr)
}

